added windows partition size support
